Код перенесен в отдельные файлы. Дорбавлена проверка на пустое значение при создании заметок и архивов.
