+++ src/main/kotlin/Element.kt
@@ -1,0 +1,1 @@
+abstract class Element(val name: String)
